+++ java-design-processor/src/test/resources/internal/nbbrd/design/NullClassNameConstant.java
@@ -1,0 +1,9 @@
+package internal.nbbrd.design;
+
+import nbbrd.design.ClassNameConstant;
+
+public class NullClassNameConstant {
+
+    @ClassNameConstant
+    public static final String ID = null;
+}